+++ CHANGELOG.md
@@ -6,6 +6,9 @@ New Features
 * Added additional before_response_sent event that gets fired even if an action was not invoked
 * Added method to add multiple custom routes at once
 
+Fixes
+------
+* Changed how routes are named to avoid collisions
 
 Version 1.1
 ===========